+++ docs/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html> <head> <meta charset="utf-8"> <title>Puestos Estaciones y Turismo</title> <meta name="viewport" content="initial-scale=1,maximum-scale=1,user-scalable=no"> <link href="https://api.tiles.mapbox.com/mapbox-gl-js/v0.46.0/mapbox-gl.css" rel="stylesheet"> <link rel="stylesheet" href="/guanacaste-map/src.f03ba8b1.css"></head> <body> <section id="main"> <div id="legend"></div> <div id="map"></div> </section> <script src="https://api.tiles.mapbox.com/mapbox-gl-js/v0.46.0/mapbox-gl.js"></script> <script src="/guanacaste-map/src.333ede2a.js"></script> </body> </html>
+<!DOCTYPE html><html> <head> <meta charset="utf-8"> <title>Puestos Estaciones y Turismo</title> <meta name="viewport" content="initial-scale=1,maximum-scale=1,user-scalable=no"> <link href="https://api.tiles.mapbox.com/mapbox-gl-js/v0.46.0/mapbox-gl.css" rel="stylesheet"> <link rel="stylesheet" href="/guanacaste-map/src.59b2d294.css"></head> <body> <section id="main"> <div id="legend"></div> <div id="map"></div> </section> <script src="https://api.tiles.mapbox.com/mapbox-gl-js/v0.46.0/mapbox-gl.js"></script> <script src="/guanacaste-map/src.6c07e544.js"></script> </body> </html>
