--- docs/src.333ede2a.js
@@ -1,14 +1,0 @@
-parcelRequire=function(e,r,n,t){var i="function"==typeof parcelRequire&&parcelRequire,o="function"==typeof require&&require;function u(n,t){if(!r[n]){if(!e[n]){var f="function"==typeof parcelRequire&&parcelRequire;if(!t&&f)return f(n,!0);if(i)return i(n,!0);if(o&&"string"==typeof n)return o(n);var c=new Error("Cannot find module '"+n+"'");throw c.code="MODULE_NOT_FOUND",c}p.resolve=function(r){return e[n][1][r]||r};var l=r[n]=new u.Module(n);e[n][0].call(l.exports,p,l,l.exports,this)}return r[n].exports;function p(e){return u(p.resolve(e))}}u.isParcelRequire=!0,u.Module=function(e){this.id=e,this.bundle=u,this.exports={}},u.modules=e,u.cache=r,u.parent=i,u.register=function(r,n){e[r]=[function(e,r){r.exports=n},{}]};for(var f=0;f<n.length;f++)u(n[f]);if(n.length){var c=u(n[n.length-1]);"object"==typeof exports&&"undefined"!=typeof module?module.exports=c:"function"==typeof define&&define.amd?define(function(){return c}):t&&(this[t]=c)}return u}({9:[function(require,module,exports) {
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=function(n){return'<div class="popup-header">\n    <h3>\n        <a\n          href="'+n.properties.link+'"\n          >\n          '+n.properties.Estación+'\n        </a>\n    </h3>\n    <div class="popup-image">\n        <a\n          href="'+n.properties.link+'"\n          >\n          <img\n            src="'+n.properties.Image+'"\n            alt="Image del Estación '+n.properties.Estación+'"\n          />\n        </a>\n      \n    </div>\n  </div>\n  <div class="popup-content">\n    <div class="popup-description">\n      <p>'+n.properties.description+'</p>\n    </div>\n    <a\n      class="popup-button"\n      href="'+n.properties.link+'"\n      >\n      Más informacíon\n    </a>\n  </div>\n  '};
-},{}],10:[function(require,module,exports) {
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.default=function(e){return'<div id="legend-inner" class="flex-child">\n\t    <h4 class="overlay-box">Puestos Estaciónes y Tourismos</h4>\n\t\t\t<div id="legend-items" class="overlay-box">\n          '+e.map(function(e){return'<div class="legend-item active" id="'+e.name+'" onClick="handleFilter(\''+e.name+"', '"+(e.type||"symbol")+"', "+(e.layerId?"'"+e.layerId+"'":null)+')">\n                  <span class="legend-key" style="background-color: '+e.color+'"></span>\n                  <span class="label">'+e.label+"</span>\n                </div>"}).join("")+'\n\n\t\t\t</div>\n      <div id="legend-footer" class="overlay-box">\n        <div onClick="noFilter()">Ver Todas las Capas</div>\n      </div>\n\t</div>'};
-},{}],8:[function(require,module,exports) {
-var o={style:"mapbox://styles/guanacaste/cjj079axn0aqu2so55fx6ln2x",center:[-85.61365526723557,10.838261234356153],zoom:9.619976883678385},e="pk.eyJ1IjoiZ3VhbmFjYXN0ZSIsImEiOiJjamowNzhuYnAwZXU2M2txczhsc21mbDVsIn0.amJMu3O1jfjcbg-B1qC7ww";module.exports={LAYER_ID:"toggle-turismo",SYMBOL_LAYERS:["toggle-turismo","toggle-ecosistemas"],ANIMATION_DURATION:2e3,LEGEND_TITLE:"Puestos Estaciónes y Tourismos",ACCESS_TOKEN:e,MAP:o,SOURCE_TYPES:[{name:"biological",layerId:"toggle-turismo",label:"Biological"},{layerId:"toggle-turismo",name:"tourist",label:"Touristo"}],LAYERS:{"toggle-turismo":{label:"Turismo",color:"#CCCC00"},"toggle-unesco":{label:"UNESCO",color:"#CCCC00"},"toggle-sectores":{label:"Sectores",color:"#449438"},"toggle-ecosistemas":{label:"Ecosistemas",color:"#000000"}}};
-},{}],6:[function(require,module,exports) {
-"use strict";Object.defineProperty(exports,"__esModule",{value:!0}),exports.getVisibleLayers=void 0;var e=require("./config");function r(e,r,t){return r in e?Object.defineProperty(e,r,{value:t,enumerable:!0,configurable:!0,writable:!0}):e[r]=t,e}var t=exports.getVisibleLayers=function(t){return t.filter(function(e){return"toggle-"===e.id.substring(0,7)}).reduce(function(t,i){var n=r({},i.id,!0);return"toggle-turismo"===i.id&&(n[i.id]=e.SOURCE_TYPES.filter(function(e){return e.layerId===i.id}).reduce(function(e,t){var i=t.name;return Object.assign(r({},i,!0),e)},{})),Object.assign(n,t)},{})};
-},{"./config":8}],7:[function(require,module,exports) {
-
-},{"./../svg/park-11.svg":[["park-11.a1b963bc.svg",19],19]}],4:[function(require,module,exports) {
-"use strict";var e=require("./components/popup"),t=a(e),n=require("./components/legend"),o=a(n),r=require("./utils");require("../assets/styles/index.scss");var l=require("./config");function a(e){return e&&e.__esModule?e:{default:e}}function i(e){if(Array.isArray(e)){for(var t=0,n=Array(e.length);t<e.length;t++)n[t]=e[t];return n}return Array.from(e)}var s={container:"map",style:"mapbox://styles/guanacaste/cjj079axn0aqu2so55fx6ln2x",center:[-85.61365526723557,10.838261234356153],zoom:9.619976883678385};mapboxgl.accessToken="pk.eyJ1IjoiZ3VhbmFjYXN0ZSIsImEiOiJjamowNzhuYnAwZXU2M2txczhsc21mbDVsIn0.amJMu3O1jfjcbg-B1qC7ww";var c=new mapboxgl.Map(Object.assign(s,l.MAP));window.map=c;var u=new mapboxgl.Popup,d=function(e){return u.setLngLat(e.geometry.coordinates).setHTML((0,t.default)(e)).addTo(c),u.on("close",function(){c.flyTo({center:s.center,duration:l.ANIMATION_DURATION})}),u};function m(){var e=document.querySelector(".mapboxgl-ctrl-bottom-left"),t=document.createElement("div");t.innerHTML='<div class="mapboxgl-ctrl mapboxgl-ctrl-group">\n      <button class="mapboxgl-ctrl-icon mapboxgl-ctrl-compass" type="button" aria-label="Reset North">\u2028\n      <span class="mapboxgl-ctrl-compass-arrow" style="transform: rotate(0deg);"></span>\u2028\n    </button>\u2028\n  </div>',t.onclick=function(){c.flyTo({center:s.center})},e.appendChild(t)}c.on("click",l.LAYER_ID,function(e){var t=e.features[0];setTimeout(function(){c.flyTo({center:t.geometry.coordinates,duration:l.ANIMATION_DURATION}),d(t)},200)}),c.addControl(new mapboxgl.FullscreenControl);var y={};c.on("load",function(){var e=new mapboxgl.NavigationControl({showCompass:!1});c.addControl(e,"top-left"),c.scrollZoom.disable(),m();var t=c.getStyle().layers;y=(0,r.getVisibleLayers)(t);var n=t.filter(function(e){var t=e.id;return"toggle-"===t.substring(0,7)&&"toggle-turismo"!==t}).map(function(e){return{name:e.id,label:l.LAYERS[e.id]?l.LAYERS[e.id].label:e.id.substring(7),type:"layer",color:l.LAYERS[e.id]?l.LAYERS[e.id].color:"darkgrey"}});document.getElementById("legend").innerHTML=(0,o.default)([].concat(i(l.SOURCE_TYPES),i(n))),window.handleFilter=function(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:null,n=arguments.length>2&&void 0!==arguments[2]?arguments[2]:null;if(document.getElementById(e).classList.toggle("active"),"symbol"===t&&n){y[n][e]=!y[n][e];var o=Object.keys(y[n]).filter(function(e){return y[n][e]});return o.length>0&&c.setLayoutProperty(n,"visibility","visible"),o.length>1?c.setFilter(n,["in","symbol"].concat(i(o))):1===o.length?c.setFilter(n,["==","symbol",o[0]]):c.setLayoutProperty(n,"visibility","none")}y[e]=!y[e];var r=y[e];c.setLayoutProperty(e,"visibility",r?"visible":"none")},window.noFilter=function(){n.map(function(e){return c.setLayoutProperty(e.name,"visibility","visible")}),n.map(function(e){var t=e.name;return c.setFilter(t,null)}),y=(0,r.getVisibleLayers)(t),document.getElementById("legend").innerHTML=(0,o.default)([].concat(i(l.SOURCE_TYPES),i(n)))}}),document.body.addEventListener("keydown",function(e){var t=e.metaKey,n=e.ctrlKey;(t||n)&&c.scrollZoom.enable()}),document.body.addEventListener("keyup",function(e){var t=e.metaKey,n=e.ctrlKey;(t||n)&&c.scrollZoom.disable()});
-},{"./components/popup":9,"./components/legend":10,"./utils":6,"../assets/styles/index.scss":7,"./config":8}]},{},[4], null)
-//# sourceMappingURL=/guanacaste-map/src.333ede2a.map
