--- docs/src.f03ba8b1.css
@@ -1,1 +1,0 @@
-#legend{min-width:240px;display:block;position:absolute;bottom:0;z-index:1;left:10%;right:10%;border-radius:3px}#legend h4{margin:0 0 2px;font-size:1.1rem}@media (min-width:460px){#legend{position:relative;max-width:240px;display:flex;flex:1;flex-direction:column;left:auto;right:auto}}.overlay-box{padding:10px;opacity:.9;margin:0 0 2px;font-size:.9em;font-family:Helvetica,Arial,Lucida Grande,sans-serif;background-color:#fff}.legend-key{display:inline-block;border-radius:50%;width:10px;height:10px;margin-right:10px}.legend-item{vertical-align:text-top}.legend-item span.label{text-transform:capitalize;display:inline-block;margin:2px;padding:2px;color:#777}.legend-item.active span.label{color:#000;font-weight:700}.legend-item:focus,.legend-item:hover{cursor:pointer}.legend-item:focus span.label,.legend-item:hover span.label{text-decoration:underline}.legend-item#tourist .legend-key{box-sizing:border-box;border:2px solid #3c565d;height:10px;width:10px;position:relative}.legend-item#biological .legend-key:before{position:relative;top:3px;border-radius:0;content:url(park-11.a1b963bc.svg);margin-left:-5px}.legend-item#toggle-ecosistemas .legend-key{border-radius:0}.legend-item#toggle-unesco .legend-key{height:3px;margin-bottom:4px;border-radius:0}.legend-item#toggle-sectores .legend-key{height:2px;margin-bottom:3px;border-radius:0}.mapboxgl-popup{width:240px;padding:0}.mapboxgl-popup-content{padding:0;box-shadow:1px 2px 2px #777}.popup-content{display:block;padding:.75rem}.popup-image{width:240px;background-color:#000}.popup-header,.popup-image{height:140px;overflow:hidden}.popup-header{display:block}.popup-header img{width:240px;height:auto;margin:-1em 0 0}.popup-header h3{display:block;position:absolute;color:#fff;margin-top:.75rem;left:.75rem;margin-right:.75rem;font-size:1.2rem;line-height:1.4rem;font-weight:400;font-family:Helvetica,Arial,Lucida Grande,sans-serif;text-shadow:1px 1px 1px #000;text-overflow:ellipsis}.popup-header h3 a{color:#fff;text-decoration:none}.popup-description{font-size:.8rem;line-height:.95rem;color:#444;max-height:120px;overflow:hidden}.popup-description p{margin:0;padding:0}.popup-button{margin-top:.5rem;display:block;text-decoration:none;font-size:.9rem;border:1px solid #05a0ce;color:#05a0ce;padding:.5rem;text-align:center;font-family:Helvetica,Arial,Lucida Grande,sans-serif}.popup-button:focus,.popup-button:hover{color:#fff;background-color:#05a0ce}.mapboxgl-popup-anchor-bottom{flex-direction:row}.mapboxgl-popup-close-button{position:absolute;right:0;top:0;border:0;border-radius:0 3px 0 0;cursor:pointer;background-color:#05a0ce69;color:#fff;font-size:1rem}.mapboxgl-popup-close-button:hover{background-color:#0a93bb69}*,body,html{margin:0;padding:0}body,html{height:100%;font-family:Helvetica,Arial,Lucida Grande,sans-serif}#main{min-height:100%;display:flex;flex-direction:row;box-sizing:border-box}#map{flex:1;flex-direction:column}
