+++ src/components/legend.js
@@ -1,12 +1,14 @@
 export default allLayers =>
   `<div id="legend-inner" class="flex-child">
-	    <h4 class="overlay-box">Puestos Estaciónes y Tourismos</h4>
-			<div id="legend-items" class="overlay-box">
+	    <h4 class="overlay-box toggle-button" onClick="tcat.toggleLegend()">Puestos Estaciónes y Tourismos</h4>
+			<div id="legend-items" class="overlay-box toggle-content">
           ${allLayers
             .map(
-              layer => `<div class="legend-item active" id="${layer.name}" onClick="handleFilter('${
+              layer => `<div class="legend-item active" id="${
                 layer.name
-              }', '${layer.type || 'symbol'}', ${layer.layerId ? `'${layer.layerId}'` : null})">
+              }" onClick="tcat.handleFilter('${layer.name}', '${layer.type || 'symbol'}', ${
+                layer.layerId ? `'${layer.layerId}'` : null
+              })">
                   <span class="legend-key" style="background-color: ${layer.color}"></span>
                   <span class="label">${layer.label}</span>
                 </div>`
@@ -14,7 +16,7 @@ export default allLayers =>
             .join('')}
 
 			</div>
-      <div id="legend-footer" class="overlay-box">
-        <div onClick="noFilter()">Ver Todas las Capas</div>
+      <div id="legend-footer" class="overlay-box toggle-content">
+        <div onClick="tcat.noFilter()">Ver Todas</div>
       </div>
 	</div>`;
